Cambio de short a Byte del atributo Qualification

+++ stamps-api/src/main/webapp/src/modules/artist/artist.mod.js
@@ -45,7 +45,7 @@ SOFTWARE.
             },
             qualification: {
                 displayName: 'Qualification',
-                type: 'short',
+                type: 'Byte',
                 required: true
             },
             popularity: {
